+++ octobot_backtesting/api/backtesting.py
@@ -13,8 +13,24 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
+from octobot_commons.constants import CONFIG_ENABLED_OPTION
 
 from octobot_backtesting.backtesting import Backtesting
+from octobot_backtesting.constants import CONFIG_BACKTESTING
+
+
+def create_backtesting(config, data_files) -> Backtesting:
+    return Backtesting(config, data_files)
+
+
+async def initialize_created_backtesting(backtesting_instance) -> Backtesting:
+    await backtesting_instance.initialize()
+    await backtesting_instance.create_importers()
+
+    if not backtesting_instance.importers:
+        raise ValueError("No importers created")
+
+    return backtesting_instance
 
 
 async def initialize_backtesting(config, data_files) -> Backtesting:
@@ -26,3 +42,22 @@ async def initialize_backtesting(config, data_files) -> Backtesting:
         raise ValueError("No importers created")
 
     return backtesting_instance
+
+
+def is_backtesting_in_progress(backtesting):
+    return backtesting.is_in_progress()
+
+
+def get_backtesting_progress(backtesting):
+    return backtesting.get_progress()
+
+
+def get_backtesting_run_report(backtesting):
+    # TODO: add backtesting report handling
+    # return backtesting.get_report()
+    return {}
+
+
+def is_backtesting_enabled(config):
+    return CONFIG_BACKTESTING in config and CONFIG_ENABLED_OPTION in config[CONFIG_BACKTESTING] \
+           and config[CONFIG_BACKTESTING][CONFIG_ENABLED_OPTION]